added dashboard layout and routing to pages from dashboard/test/leaderboard/index.tsx
